moving the USSDMenu code for the screens and the services
