+++ examples/makefile
@@ -1,0 +1,2 @@
+run:
+	cargo run --example basic_usage
