+++ src/prelude.rs
@@ -1,2 +1,3 @@
 pub use crate::menu::MenuBuilder;
 pub use crate::UssdApp;
+pub use crate::USSDMenu
